backend/requirements.txt
# Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
python-multipart==0.0.6
httpx==0.26.0
openai==1.12.0
elevenlabs==1.1.2
python-google-generativeai==0.3.2
ffmpeg-python==0.2.0
moviepy==1.0.3
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2024.1
uuid==1.30

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2
